% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_set_active_touchstone}
\alias{admin_set_active_touchstone}
\title{Set the active touchstone}
\usage{
admin_set_active_touchstone(con, touchstone_id, dry_run = TRUE)
}
\arguments{
\item{con}{Database connection (will require write access to the database)}

\item{touchstone_id}{Touchstone id to set as "open"}

\item{dry_run}{Don't commit the transaction - just test if it would work}
}
\description{
Set the active touchstone
}
