% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_burden_templates.R
\name{create_burden_template}
\alias{create_burden_template}
\title{Creat burden estimate templates}
\usage{
create_burden_template(files, files_path,
  assert_files = c("Country_Disease_Table.csv", "model_meta.csv",
  "model_outcomes.csv"), central_templates_only = FALSE)
}
\arguments{
\item{files}{filenames, it has to contain the following three files
"Country_Disease_Table.csv", "model_meta.csv" and "model_outcomes.csv"}

\item{files_path}{the directory that contains those files above}

\item{assert_files}{needed meta files
"Country_Disease_Table.csv", "model_meta.csv" and "model_outcomes.csv"}

\item{central_templates_only}{TRUE if you only need central burden templates, usually for open call. But it depend on Tini's input.
Set to FALSE as default, which outputs both central and stochastic templates}
}
\description{
Creat burden estimate templates
}
