% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_touchstone.R
\name{create_touchstone}
\alias{create_touchstone}
\title{Create a new touchstone}
\usage{
create_touchstone(con, dat, demography_from = NULL, path_meta = "meta",
  transaction = TRUE, dry_run = TRUE)
}
\arguments{
\item{con}{Database connection.  You will need to be the
\code{vimc} or \code{import} user (not \code{readonly}) to run
this function.}

\item{dat}{A \code{data.frame} of coverage data to import.}

\item{demography_from}{Touchstone id to import
demographic statistics from.  This must currently be given, but
in future we'll allow this to be imported from a csv}

\item{path_meta}{Path that we look for various metadata files.
Eventually we'll document what they look like.}

\item{transaction}{}

\item{dry_run}{}
}
\description{
Create and import a new touchstone.
}
