% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database_connection}
\alias{database_connection}
\title{Connect to database}
\usage{
database_connection(location = "science", user = "readonly",
  local_port = NULL, local_password_group = "science")
}
\arguments{
\item{location}{One of "science", "production", "uat" or
"localhost".  Be \emph{very} careful if using production.  If
\code{NULL}, then the location is set by the
\code{MONTAGU_DB_HOST} environment variable.  For running in
import scripts you can arrange to set that to the value of the
shell command \code{hostname} (e.g. \code{export
MONTAGU_DB_HOST=$(hostname)}).}

\item{user}{Username to connect as}

\item{local_port}{Port (when running locally)}

\item{local_password_group}{Password group (when running locally)}
}
\description{
Connect to database
}
