% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{fix_coverage_fvps}
\alias{fix_coverage_fvps}
\title{Impact Calculation (method 2)}
\usage{
fix_coverage_fvps(con, touchstone_name = "201710gavi", year_min = 2000,
  year_max = 2100, pine = FALSE, write_table = TRUE,
  report_suspecious_coverage = FALSE, touchstone_pop = NULL,
  gavi_support_levels = c("with", "bestminus"))
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{touchstone_name}{Specify touchstone name only, not with specific version.}

\item{year_min}{min year of vaccination}

\item{year_max}{max year of vaccination}

\item{pine}{this is for testthat. we only grab data for pine countries if true}

\item{write_table}{If true, create a temporary table; otherwise return a dataframe}

\item{report_suspecious_coverage}{switch on/off the reporting of suspecious coverage}

\item{touchstone_pop}{population touchstone, this is for the modups
where fvps are calculated using a coverage touchstone and a population touchstone}
}
\description{
Provide age-specific coverage-un_pop-fvps
}
