% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_update.R
\name{modified_update_calculate}
\alias{modified_update_calculate}
\title{Do a modified update}
\usage{
modified_update_calculate(con, touchstone_name_mod, touchstone_use,
  method = 2, version = 1)
}
\arguments{
\item{con}{Database connection}

\item{touchstone_name_mod}{Name of the new modified touchstone
that we are creating}

\item{touchstone_use}{Name of the touchstone that we are basing
this off of}

\item{method}{two options: 1 = method1, 2 = method2
method2_v2 is different from method2 in that, it takes per-year-gavi-level as an imput for gavi impact}

\item{version}{two options: 1 = v1 and 2 = v2}
}
\description{
Do a modified update
}
