% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_update.R
\name{modified_update_summary_output}
\alias{modified_update_summary_output}
\title{Modified update summary output}
\usage{
modified_update_summary_output(con, res, path_meta)
}
\arguments{
\item{con}{Database connection}

\item{res}{A modified update object (returned from
\code{modified_update_calculate})}

\item{path_meta}{Path to the metadata directory.  In this
directory the following files must exist:
\code{gavi_country_data.csv}, \code{tr_touchstone.csv},
and \code{years_output.csv}.  There is no
validation done on these files and providing the wrong thing
will break in interesting and unknown ways.}
}
\description{
Modified update summary output
}
