% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_update.R
\name{mu_year_introduction}
\alias{mu_year_introduction}
\title{Find year of introduction}
\usage{
mu_year_introduction(con, dat, dat_summary)
}
\arguments{
\item{con}{Database connection}

\item{dat}{Data: the list output from the modup}

\item{dat_summary}{Data: the summary output from the modup}
}
\description{
Look for introduction year and add to summary output
}
