% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{prepare_recipe}
\alias{prepare_recipe}
\title{Prepare impact recipe}
\usage{
prepare_recipe(con, recipe = "impact.csv")
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{recipe}{at the moment, it is a csv file.
Once imported, recipe will be more flexible}
}
\description{
Prepare impact calculation recipes
}
