% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_touchstone.R
\name{project_coverage}
\alias{project_coverage}
\title{Project coverage}
\usage{
project_coverage(dat, year_project_from, year_from = 1980, year_to = 2100)
}
\arguments{
\item{dat}{Data with columns...}

\item{year_project_from}{This is the year that the projections start}

\item{year_from, year_to}{Range of the actual coverage data that
that you want.}
}
\description{
Project coverage
}
