% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{test_data}
\alias{test_data}
\title{Generate test data}
\usage{
test_data(con, con_test, modelling_group = "PSU-Ferrari", year_min = 2000,
  year_max = 2030)
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{con_test}{This is a SQLite connection}

\item{modelling_group}{this is the modelling_group that the test data is linked to}

\item{year_min}{min year of vaccination}

\item{year_max}{max year of vaccination}
}
\description{
Make a small chunck of test data for testing purpose
}
