% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalys.R
\name{calculate_dalys}
\alias{calculate_dalys}
\title{DALYs calculation}
\usage{
calculate_dalys(con, touchstone_name, year_min = 2000, year_max = 2030,
  vimc_dalys_only = TRUE, modelling_group = NULL)
}
\arguments{
\item{con}{You can be \code{readonly} user to run this function.
But if you need to import dalys for Ferrari, Li and LiST (201710gavi), you can use \code{import} user.}

\item{touchstone_name}{touchstone for which dalys are calcualted}

\item{year_min}{minimal year}

\item{year_max}{maximum year}

\item{vimc_dalys_only}{set to be TRUE if we are only interested in Ferrari, Li and LiST (201710gavi)}

\item{modelling_group}{This parameter makes the calcualtion more flexible, specify a vector of modelling_group(s) that you are interested in.}
}
\description{
Calculate dalys
}
