% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database_connection}
\alias{database_connection}
\title{Connect to database}
\usage{
database_connection(
  location = "science",
  user = "readonly",
  local_port = NULL,
  local_password_group = "science"
)
}
\arguments{
\item{location}{One of "science", "production", "uat" or
"localhost".  Be \emph{very} careful if using production}

\item{user}{Username to connect as}

\item{local_port}{Port (when running locally)}

\item{local_password_group}{Password group (when running locally)}
}
\description{
Connect to database
}
