% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{fix_coverage_fvps}
\alias{fix_coverage_fvps}
\title{Impact Calculation (method 2)}
\usage{
fix_coverage_fvps(con, touchstone_name = "201710gavi")
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{touchstone_name}{}
}
\description{
Provide age-specific coverage-un_pop-fvps
}
