% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{impact_calculation}
\alias{impact_calculation}
\title{Impact Calculation (method 2)}
\usage{
impact_calculation(con, touchstone_name = "201710gavi", cohort_min = 2000,
  cohort_max = 2030, routine_tot_rate_shape = "trace_cohort",
  method = "method2")
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{touchstone_name}{is the touchstone relevant to which impact is calculated}

\item{cohort_min}{minimal birth year of interest}

\item{cohort_max}{maximal birth year year of interest}

\item{routine_tot_rate_shape}{This parameter determines how we chop off the year-age matrix to calculate impact rates
campaign is stratiforward, use all fvps and all burden estimates to calculate impact rate. So no need to specify.
Becuase all impacts (years 2000-2100) are derived from campaigns between 2000 and 2030.
Routine is more complicated. We either trance birth cohort between 2000-2030 or trance all birth cohorts between 2000-2100.}

\item{method}{impact calculation method - chose from method1 and method2
impact outcome can be provided as age specific if simplified=FALSE}
}
\description{
Calcualte impact with method2 that allocate impact by fvps*impact_rate
}
