% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_method2.R
\name{impact_calculation}
\alias{impact_calculation}
\title{Impact Calculation (method 2)}
\usage{
impact_calculation(con, touchstone_name = "201710gavi", age_min = 0,
  age_max = 100, year_min = 2000, year_max = 2030, method = "method2")
}
\arguments{
\item{con}{Database connection.  You will need to be \code{readonly} user
to run this function.}

\item{touchstone_name}{is the touchstone relevant to which impact is calculated}

\item{age_min}{minimal age of interest}

\item{age_max}{maximal age of interest}

\item{year_min}{minimal year of interest}

\item{year_max}{maximal year of interest}

\item{simplified}{if simplified=TRUE, impact outcome is presented as for per year
impact outcome can be provided as age specific if simplified=FALSE}
}
\description{
Calcualte impact with method2 that allocate impact by fvps*impact_rate
}
