% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_update.R
\name{modified_update_calculate}
\alias{modified_update_calculate}
\title{Do a modified update}
\usage{
modified_update_calculate(con, touchstone_name_mod, touchstone_use)
}
\arguments{
\item{con}{Database connection}

\item{touchstone_name_mod}{Name of the new modified touchstone
that we are creating}

\item{touchstone_use}{Name of the touchstone that we are basing
this off of}
}
\description{
Do a modified update
}
